requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-extensions==0.7.1
drf-nested-routers==0.93.5
Markdown==3.6
psycopg2-binary==2.9.9
PyJWT==2.8.0
sqlparse==0.4.4
